Complete Create inspection workflow done - All data available in last function just need to send via API to the backend Nousheen you have to work on this.
1 -  MOdal added for camera & gallery
2 - Signature pad added
3 -Bug fixes and refactoring
4 -  Filtering and search done

Nousheen you have to work on create question API as well. Than 90 functional work done minor bugs and tweaks.

Ask some UI developer to work on it.

PLease write clean code and maintain it. I am relying on you for rest of the items as all major ones are completed.